refactor: 클래스 패키지 구조 수정
