--- src/main/kotlin/minesweeper/Mines.kt
@@ -1,6 +1,0 @@
-package minesweeper
-
-@JvmInline
-value class Mines(
-    val mines: Set<Position>
-)
