+++ src/main/kotlin/minesweeper/board/GameBoardRenderStrategy.kt
@@ -1,4 +1,7 @@
-package minesweeper
+package minesweeper.board
+
+import minesweeper.board.BoardDimensions
+import minesweeper.board.RenderedGameBoard
 
 fun interface GameBoardRenderStrategy {
