+++ src/main/kotlin/minesweeper/exception/BoardSizeOverException.kt
@@ -1,3 +1,3 @@
-package exception
+package minesweeper.exception
 
 class BoardSizeOverException(message: String): Exception(message)
