+++ src/main/kotlin/minesweeper/mine/MineCount.kt
@@ -1,11 +1,10 @@
-package minesweeper
+package minesweeper.mine
 
 @JvmInline
-value class Height(
-    val value: Int
+value class MineCount(
+    private val value: Int
 ) {
     constructor(input: String) : this(input.toInt())
 
-    operator fun minus(other: Int) = value - other
     operator fun compareTo(other: Int) = this.value - other
 }
