+++ src/main/kotlin/minesweeper/mine/Mines.kt
@@ -1,0 +1,8 @@
+package minesweeper.mine
+
+import minesweeper.position.Position
+
+@JvmInline
+value class Mines(
+    val mines: Set<Position>
+)
