+++ src/main/kotlin/minesweeper/position/Position.kt
@@ -1,4 +1,4 @@
-package minesweeper
+package minesweeper.position
 
 data class Position(
     val y: Int,
