+++ src/main/kotlin/minesweeper/position/PositionGenerateStrategy.kt
@@ -1,4 +1,4 @@
-package minesweeper
+package minesweeper.position
 
 fun interface PositionGenerateStrategy {
     fun generate(): Position
